Extract `FlightTag` component from `Version` and `DeletionStatus`
