+++ src/routes/organizations/routes/aircraft/components/FlightList/DeletionStatus.js
@@ -3,19 +3,10 @@ import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import { FormattedMessage } from 'react-intl'
 import { withStyles } from '@material-ui/core'
-import Typography from '@material-ui/core/Typography'
 import { flight as flightShape } from '../../../../../../shapes'
+import FlightTag from './FlightTag'
 
 const styles = theme => ({
-  status: {
-    textTransform: 'uppercase',
-    backgroundColor: theme.palette.grey[200],
-    borderRadius: theme.typography.pxToRem(5),
-    padding: theme.typography.pxToRem(5),
-    display: 'inline-block',
-    verticalAlign: 'top',
-    fontSize: '0.8em'
-  },
   deleted: {
     backgroundColor: theme.palette.error.dark,
     color: '#fff'
@@ -27,10 +18,8 @@ const styles = theme => ({
 })
 
 const DeletionStatus = ({ flight, classes, className }) => (
-  <Typography
-    component="span"
+  <FlightTag
     className={classNames(
-      classes.status,
       flight.replacedWith ? classes.replaced : classes.deleted,
       className
     )}
@@ -38,7 +27,7 @@ const DeletionStatus = ({ flight, classes, className }) => (
     <FormattedMessage
       id={`flightlist.flight.${flight.replacedWith ? 'replaced' : 'deleted'}`}
     />
-  </Typography>
+  </FlightTag>
 )
 
 DeletionStatus.propTypes = {
