+++ src/routes/organizations/routes/aircraft/components/FlightList/Version.js
@@ -1,43 +1,27 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import classNames from 'classnames'
 import { FormattedMessage } from 'react-intl'
 import { withStyles } from '@material-ui/core'
-import Typography from '@material-ui/core/Typography'
-import { flight } from '../../../../../../shapes'
+import { flight as flightShape } from '../../../../../../shapes'
+import FlightTag from './FlightTag'
 
-const styles = theme => ({
-  version: {
-    textTransform: 'uppercase',
-    backgroundColor: theme.palette.grey[200],
-    borderRadius: theme.typography.pxToRem(5),
-    padding: theme.typography.pxToRem(5),
-    display: 'inline-block',
-    verticalAlign: 'top',
-    fontSize: '0.8em'
-  },
+const styles = {
   active: {
     backgroundColor: '#81c784'
   }
-})
+}
 
 const Version = ({ flight, classes }) => (
-  <Typography
-    component="span"
-    className={classNames(
-      classes.version,
-      !flight.replacedWith && classes.active
-    )}
-  >
+  <FlightTag className={!flight.replacedWith ? classes.active : null}>
     <FormattedMessage
       id="flightlist.flight.version"
       values={{ version: flight.version }}
     />
-  </Typography>
+  </FlightTag>
 )
 
 Version.propTypes = {
-  flight: flight.isRequired,
+  flight: flightShape.isRequired,
   classes: PropTypes.object.isRequired
 }
